requirements.txt
annotated-types==0.7.0
asn1crypto==1.5.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
Faker==27.0.0
filelock==3.15.4
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
jaraco.classes==3.4.0
Jinja2==3.1.4
keyring==24.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.4.0
packaging==24.1
platformdirs==4.2.2
pluggy==1.5.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requirements-parser==0.9.0
rich==13.7.1
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
snowflake-cli-labs==2.7.0
snowflake-connector-python==3.11.0
sortedcontainers==2.4.0
toml==0.10.2
tomlkit==0.13.0
typer==0.12.3
types-setuptools==71.1.0.20240806
typing_extensions==4.12.2
urllib3==2.2.2
